aggiunti ulteriori servizi di controllo per le classi che estendono il poi

+++ api/src/main/java/it/unicam/cs/model/DTO/MonumentoDto.java
@@ -12,7 +12,7 @@ public class MonumentoDto extends PoiDto{
     private int annoRealizzazione;
     private String descrizione;
     private String autore;
-    private int altezza;
-    private int lunghezza;
+    private double altezza;
+    private double lunghezza;
     private String architettura;
 }
